pyproject.toml
[project]
name = "helena-agent"
version = "0.1.0"
description = "LangGraph multi-agent service for Helena security operations"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "langgraph>=1.1",
    "langgraph-checkpoint-postgres>=2.0",
    "openai>=1.35",
    "asyncpg>=0.29",
    "langfuse>=2.20",
    "structlog>=24.1",
    "prometheus-client>=0.20",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "sse-starlette>=2.1",
    "psycopg[binary]>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pytest-mock>=3.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
]
